Renamed CommandType to CommandOperation
